+++ src/dom.js
@@ -1,5 +1,4 @@
 const dom = {
     container: document.querySelector('.weather-api')
 };
-
 export default dom;
