--- src/load-event.js
@@ -1,5 +1,0 @@
-import loadHandler from './loadHandler.js';
-
-const loadEvent = window.addEventListener('load', loadHandler);
-
-export default loadEvent;
